Add tool refs within exports
