+++ app-backend/tool/src/main/scala/ast/package.scala
@@ -1,12 +1,45 @@
 package com.azavea.rf.tool
 
-import java.util.UUID
-
+import com.azavea.rf.tool.ast._
+import com.azavea.rf.tool.ast.MapAlgebraAST._
 import com.azavea.rf.tool.ast.codec.MapAlgebraCodec
+
 import io.circe._
 import io.circe.optics.JsonPath._
+import cats._
+import cats.data._
+import cats.implicits._
+
+import scala.concurrent._
+import java.util.UUID
 
 package object ast extends MapAlgebraCodec {
+
+  /** Collect resources necessary to carry out substitutions on  ASTs while avoiding cycles */
+  def assembleSubstitutions(
+    ast: MapAlgebraAST,
+    substitutionResolver: UUID => Future[Option[MapAlgebraAST]],
+    assembled: Map[UUID, MapAlgebraAST] = Map()
+  )(implicit ec: ExecutionContext): OptionT[Future, Map[UUID, MapAlgebraAST]] =
+    ast match {
+      case ToolReference(id, refId) =>
+        OptionT(substitutionResolver(refId)).flatMap({ astPatch =>
+          // Here's where we can hope to catch cycles as we traverse down
+          //  the tree (keeping track of previously encountered references)
+          if (assembled.get(refId).isDefined)
+            // TODO: EitherT this thing
+            throw new IllegalArgumentException(s"Repeated substitution $refId found; likely cyclical reference")
+          else
+            assembleSubstitutions(astPatch, substitutionResolver, assembled ++ Map(refId -> astPatch))
+        })
+      case op: MapAlgebraAST.Operation =>
+        val childSubstitutions = op.args.map({ arg => assembleSubstitutions(arg, substitutionResolver, assembled) }).sequence
+        childSubstitutions.map({ substitutions =>
+          substitutions.foldLeft(Map[UUID, MapAlgebraAST]())({ case (map, subs) => map ++ subs }) ++ assembled
+        })
+      case _ =>
+        OptionT.pure[Future, Map[UUID, MapAlgebraAST]](assembled)
+      }
 
   implicit class CirceMapAlgebraJsonMethods(val self: Json) {
     def _id: Option[UUID] = root.id.string.getOption(self).map(UUID.fromString(_))
